requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
torch  # Install PyTorch with CUDA if necessary (use the correct version as per your CUDA version)
ffmpeg-python  # Python bindings for FFmpeg
transformers
openai

# Whisper from the official GitHub repository
git+https://github.com/openai/whisper.git